Performance: improve .trim performance for large strings contains lots of whitespaces
The old implementation took O(N^2) time to trim the string when multiple adjacent spaces
were present.

For instance, consider the string "A          B" (10 spaces between A and B).
Then old regexp /[\s]+$/ would take 10 steps until it realizes
the regexp does not match at position 1.
Then it would try to match the regexp at position 2, and it would take 9 steps.
Then 8 steps for position 3, ... so it would take 10*10/2 steps until
it figures out the regexp does not match.

The new approach is to require "non-whitespace" char before the whitespace run,
so it spends just one step for each space in the string.